deleted float16 converison for precison

+++ flash_attention/program.py
@@ -328,7 +328,7 @@ def _attn_bwd_dk_dv(
         dV_block += tl.dot(pT, dO_block.to(tl.float32)).to(tl.float32)
         D_block = tl.load(D+offsets_q)
         dpT = tl.dot(V_block, tl.trans(dO_block)).to(tl.float32)
-        dS_T = pT * (dpT - D_block).to(tl.float16)
+        dS_T = pT * (dpT - D_block)
         dK_block += softmax_scale * tl.dot(dS_T, tl.trans(qT_block).to(tl.float32)) # float32?
